analytics/requirements_base.txt
clickhouse-connect==0.8.12
git+https://github.com/netbek/clickhouse-sqlalchemy.git@datetime-uuid
dbt-artifacts-parser==0.8.1
git+https://github.com/netbek/dbt-clickhouse.git@v1.8.7-batch
dbt-core==1.8.9  # Pinned because of dbt-clickhouse and prefect-dbt
httpx==0.28.1
pandas==2.2.3
prefect==3.0.10
prefect-dbt==0.6.4
prefect-jupyter==0.3.1
prefect-shell==0.3.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
pydash==8.0.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==13.9.4
sqlglot==26.0.1
sqlparse==0.5.3
typer==0.12.5  # Pinned because of prefect
